backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
pandas
numpy
scikit-learn
langdetect
deep-translator
google-generativeai
groq
reportlab
python-dotenv
aiofiles
sse-starlette
python-jose[cryptography]
passlib[bcrypt]
httpx
